requirements.txt
cryptography==36.0.2
Flask==2.1.0
matplotlib==3.3.4
names==0.3.0
numpy==1.20.1
pandas==1.2.3
Pillow==9.0.1
pydicom==2.2.2
requests==2.25.1
rstr==3.1.0
